backend/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
camelot-py==1.0.0
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.1.8
cryptography==45.0.6
et_xmlfile==2.0.0
exceptiongroup==1.3.0
fastapi==0.116.1
ghostscript==0.8.1
h11==0.16.0
idna==3.10
numpy==2.0.2
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
pandas==2.3.1
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pypdf==3.17.4
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
six==1.17.0
sniffio==1.3.1
starlette==0.47.2
tabulate==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uvicorn==0.35.0
